+++ histou/grafana/graphpanel/graphpanelinfluxdb.php
@@ -55,6 +55,9 @@ class GraphPanelInfluxdb extends GraphPanel
         $tags = array();
         $i = 0;
         foreach ($filterTags as $key => $value) {
+            if($value === NULL) {
+                continue;
+            }
             $condition = (array_key_exists('condition', $value) ? $value['condition'] : 'AND');
             $operator = (array_key_exists('operator', $value) ? $value['operator'] : '=');
             if($operator === '=' && preg_match('/^\/.*\/$/', $value['value'])) {
@@ -90,14 +93,14 @@ class GraphPanelInfluxdb extends GraphPanel
             $target = $this->createTarget(array(
                                             'host' => array('value' => \histou\helper\str::genRegex($host), 'operator' => '=~'),
                                             'service' => array('value' => \histou\helper\str::genRegex($service), 'operator' => '=~'),
-                                            'command' => array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
+                                            'command' => $command == NULL ? NULL : array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
                                             'performanceLabel' => array('value' => \histou\helper\str::genRegex($performanceLabel), 'operator' => '=~')
                                             ));
         } else {
             $target = $this->createTarget(array(
                                             'host' => array('value' => $host),
                                             'service' => array('value' => $service),
-                                            'command' => array('value' => $command),
+                                            'command' => $command == NULL ? NULL : array('value' => $command),
                                             'performanceLabel' => array('value' => $performanceLabel)
                                             ));
         }
@@ -171,7 +174,7 @@ class GraphPanelInfluxdb extends GraphPanel
                 array(
                         'host' => array('value' => \histou\helper\str::genRegex($host), 'operator' => '=~'),
                         'service' => array('value' => \histou\helper\str::genRegex($service), 'operator' => '=~'),
-                        'command' => array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
+                        'command' => $command == NULL ? NULL : array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
                         'performanceLabel' => array('value' => \histou\helper\str::genRegex($performanceLabel), 'operator' => '=~'),
                         'downtime' => array('value' => "true")
                     )
@@ -181,7 +184,7 @@ class GraphPanelInfluxdb extends GraphPanel
                 array(
                         'host' => array('value' => $host),
                         'service' => array('value' => $service),
-                        'command' => array('value' => $command),
+                        'command' => $command == NULL ? NULL : array('value' => $command),
                         'performanceLabel' => array('value' => $performanceLabel),
                         'downtime' => array('value' => "true")
                     )
@@ -243,7 +246,7 @@ class GraphPanelInfluxdb extends GraphPanel
                 array(
                         'host' => array('value' => \histou\helper\str::genRegex($host), 'operator' => '=~'),
                         'service' => array('value' => \histou\helper\str::genRegex($service), 'operator' => '=~'),
-                        //'command' => array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
+                        //'command' => $command == NULL ? NULL : array('value' => \histou\helper\str::genRegex($command), 'operator' => '=~'),
                         'performanceLabel' => array('value' => \histou\helper\str::genRegex($performanceLabel), 'operator' => '=~'),
                     ),
                 FORECAST_DATASOURCE_NAME
@@ -253,7 +256,7 @@ class GraphPanelInfluxdb extends GraphPanel
                 array(
                         'host' => array('value' => $host),
                         'service' => array('value' => $service),
-                        //'command' => array('value' => $command),
+                        //'command' => $command == NULL ? NULL : array('value' => $command),
                         'performanceLabel' => array('value' => $performanceLabel),
                     ),
                 FORECAST_DATASOURCE_NAME